+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## New Features
 
+- Add `scripts/sl.cmake` to install sl and a clean compiler
+- Add `scripts/profile.sh` to set the ENV variables from a project
 - add `sos_sdk_copy_file` to copy a file without overwriting
 - add `sos_sdk_overwrite_file` to copy a file with overwriting
 - add `sos_sdk_git_clone_or_pull` to pull subprojects